requirements_analysis.txt
# Core scientific computing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# MCMC and statistics
pystan>=3.0.0
arviz>=0.11.0
emcee>=3.1.0
corner>=2.2.0

# Alternative MCMC (choose one)
# numpyro>=0.8.0
# jax>=0.2.0
# jaxlib>=0.1.0

# Cosmology
astropy>=4.3.0
# camb>=1.3.0  # Optional: for generating theory spectra
# cobaya>=3.2.0  # Optional: for cosmological MCMC

# Plotting
matplotlib>=3.4.0
seaborn>=0.11.0

# Data formats
pyyaml>=5.4.0
h5py>=3.0.0

# Testing and validation
pytest>=6.2.0
pytest-cov>=2.12.0

# Development tools
black>=21.0
flake8>=3.9.0
mypy>=0.900

# Documentation
sphinx>=4.0.0
sphinx-rtd-theme>=0.5.0

# Notebooks (optional)
# jupyter>=1.0.0
# jupyterlab>=3.0.0
# ipywidgets>=7.6.0